+++ ozygod-pm-model/src/main/java/com/ozygod/model/zdconfig/vo/game/JDSLGameConfigVo.java
@@ -3,6 +3,7 @@ package com.ozygod.model.zdconfig.vo.game;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -25,7 +26,7 @@ public class JDSLGameConfigVo extends BaseGameConfigVo {
 //     * join_delay : {"min":1,"max":10}
 //     * num : {"min":13,"max":13}
      */
-    private int stragy;
+    private BigDecimal stragy;
     private MinMaxVo time;
     private MinMaxVo max_gold;
     private String logic;
